Issue 101: mbus-master: Add a flag to skip the pings.
The rationale for the `init_slaves` call seems undocumented (it was
added when the `mbus-{tcp,serial}-request-data-multi-reply.c` files were
added to `libmbus`).  If the slaves do not respond to the ping
operation, then it needlessly wastes double whatever the time-out value
is set to.

The ping immediately before communicating with a slave by primary
address seems to be a work-around to a specific model of M-Bus meter:
the Sontex Supercal531.

https://github.com/rscada/libmbus/pull/95

This adds a boolean flag that permits enabling or disabling this
work-around as needed.