Issue 101: index.js: Expose some options for `getData`
We'll use an object to allow multiple options to be specified in a
single object (perhaps the same object if the caller desires).  If the
`options` parameter winds up being a call-back function, then assume no
options are set.

This maintains backward-compatibility with existing scripts whilst still
allowing for future expansion.

Finally, document the `pingFirst` (our one and only option so far).